+++ RELEASENOTES.md
@@ -1,5 +1,27 @@
 These are the release notes for the [wikibase-api](README.md).
 
+## Version 0.3 (in development)
+
+####Fixes
+
+* You can now clear an entity using RevisionSaver by passing an empty Item/Property/Entity object.
+  Prior to this a UsageException would be thrown as the clear param was not passed to the api.
+* SiteLinkSetter now uses the correct API module (wbsetsitelink).
+  Previous to this it was for some reason trying to use 'wblinktitles'.
+
+####Changes
+
+* RevisionSaver->save() now returns an Entity object instead of always returning true.
+
+####Libs
+
+We now require the following:
+* "wikibase/data-model-serialization": "~1.0.1"
+* "wikibase/data-model": "~0.8.2"
+* "data-values/geo": "~0.2.0"
+* "data-values/data-types": "~0.4.0",
+
+
 ## Version 0.2 (2014-07-15)
 
 * Adjust for mediawiki-api 0.3